model/requirements.txt
cmdstanpy==0.9.68
convertdate==2.3.2
cycler==0.10.0
Cython==0.29.23
ephem==4.0.0.2
grpcio==1.41.1
grpcio-tools==1.41.1
hijri-converter==2.1.3
holidays==0.11.1
kiwisolver==1.3.1
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
matplotlib==3.4.2
numpy==1.21.0
pandas==1.3.0
Pillow==8.3.2
protobuf==3.19.1
PyMeeus==0.5.11
pyparsing==2.4.7
python-dateutil==2.8.1
python-json-logger==2.0.2
pytz==2021.1
setuptools-git==1.2
six==1.16.0
tqdm==4.61.1
ujson==4.0.2
